+++ testsuite/ada_lsp/find_all_refs_extending/common.adb
@@ -1,0 +1,13 @@
+with Ada.Text_IO;
+
+package body Common is
+
+   procedure Put (Item : Enum := Hello) is
+   begin
+      case Item is
+         when Hello => null;
+         when Goodbye => null;
+      end case;
+   end Put;
+
+end Common;