+++ testsuite/ada_lsp/find_all_refs_extending/common/common.ads
@@ -1,0 +1,7 @@
+package Common is
+
+   type Enum is (Hello, Goodbye);
+
+   procedure Put (Item : Enum := Hello);
+
+end Common;