+++ testsuite/ada_lsp/find_all_refs_extending/program/program.adb
@@ -1,0 +1,9 @@
+with Common;
+
+procedure Program is
+begin
+
+   Common.Put (Common.Hello);
+   Common.Put (Common.Goodbye);
+
+end Program;